+++ src/pages/Schedule.tsx
@@ -3,7 +3,7 @@ import { useState } from 'react';
 import { css } from '@emotion/react';
 
 import CalendarComponent from '@/components/Calendar/Calendar';
-import DaySchedule from '@/components/daySchedule/DaySchedule';
+import DailySchedule from '@/components/dailySchedule/DailySchedule';
 import useFetchSchedule from '@/hooks/useFetchSchedule';
 import theme from '@/styles/theme';
 
@@ -18,7 +18,7 @@ const SchedulePage = () => {
         <CalendarComponent events={events} activeDate={activeDate} setActiveDate={setActiveDate} />
       </div>
       <div>
-        <DaySchedule date={activeDate.toISOString()} schedules={schedule} />
+        <DailySchedule date={activeDate.toISOString()} schedules={schedule} />
       </div>
     </>
   );
